+++ docs/release-notes.md
@@ -35,6 +35,8 @@ of `zizmor`.
   block-style YAML mappings (#995)
 * Fixed a bug where `zizmor` would crash when given a GitHub API token
   with leading or trailing whitespace (#1027)
+* Fixed a bug where [template-injection] findings in `--fix` mode would be
+  incorrectly patched when referencing an `env.*` context (#1052)
 
 ## 1.11.0